+++ src/components/dash/OrgAnnotation.vue
@@ -1,28 +1,31 @@
 <template>
   <div v-if="!editable">
     <span v-if="value.tag !== 'inhibition'" class="label label-info"
+      style="display:box"
       >{{ value.tag }}</span>
-    <div v-else class="list-group-item-warning">
-      <div class="col-sm-2">Inhibition:
+    <div v-else class="list-group-item list-group-item-warning row">
+      <div>Inhibition:
       </div>
       <annotation-condition class="col-sm-10"
         v-model="value.condition" v-bind:status="status"
         v-bind:condition-hints="conditionHints"/>
     </div>
   </div>
-  <div v-else class="list-group form-horizontal">
+  <div v-else class="list-group form-inline">
     <div v-bind:class="annoClass">
-      <div v-if="value.tag === 'inhibition'" class="form-group">
+      <div v-if="value.tag === 'inhibition'" class="form-group-sm">
         Inhibition:
         <annotation-condition v-model="value.condition"
           v-bind:status="status"
           v-bind:condition-hints="conditionHints"/>
       </div>
-      <div v-else class="form-group">
-        <label class="col-sm-3 control-label">Tag</label>
-        <div class="col-sm-8">
-          <input type="text" v-model="value.tag" class="form-control" />
-        </div>
+      <div v-else class="form-group-sm">
+        <label class="control-label">Tag</label>
+        <select v-model='selectionValue' class="form-control btn-info">
+          <option disabled value="">(custom)</option>
+          <option v-for="tag in annotationHints.tags">{{ tag }}</option>
+        </select>
+        <input type="text" v-model="value.tag" class="form-control" />
       </div>
     </div>
     <button v-on:click="$emit('deleteMe')" class="btn btn-default btn-xs">
@@ -47,6 +50,7 @@ module.exports = {
   },
   data: function () {
     return {
+      selectionValue: ''  // value of tag's <select>
     }
   },
   components: {
@@ -66,10 +70,38 @@ module.exports = {
     annoClass: function () {
       return {
         'list-group-item': this.value.tag !== 'inhibition',
-        'list-group-item list-group-item-warning':
+        'list-group-item list-group-item-warning row':
           this.value.tag === 'inhibition'
       }
     }
+  },
+  watch: {
+    // we want to allow value.tag to have either a value from the
+    // <select> options from annotationHints.tags or a custom,
+    // so we have to use a second variable 'selectionValue' and functions
+    // to implement that behaviour via watching it an value.tag:
+    selectionValue: function (newSelectionValue) {
+      if (newSelectionValue !== '') {
+        this.value.tag = newSelectionValue
+      }
+    },
+    value: {
+      handler: function (newValue) { this.updateSelection(newValue) },
+      deep: true
+    }
+  },
+  methods: {
+    updateSelection: function (newValue) {
+      if (this.annotationHints.tags.indexOf(newValue.tag) >= 0) {
+        this.selectionValue = newValue.tag
+      } else {
+        this.selectionValue = ''
+      }
+    }
+  },
+  mounted: function () {
+    // make sure the select choice reflects the initial value.tag.
+    this.updateSelection(this.value)
   }
 }
 </script>